reorganize utils and avoid running multiple jvms
